+++ crates/bin/c8y_configuration_poc/src/c8y_http_proxy/actor.rs
@@ -38,7 +38,7 @@ use tedge_http_ext::HttpRequest;
 use tedge_http_ext::HttpRequestBuilder;
 use tedge_http_ext::HttpResponseExt;
 use tedge_http_ext::HttpResult;
-use url::Url;
+use tedge_utils::file::move_file;
 
 const RETRY_TIMEOUT_SECS: u64 = 60;
 
@@ -167,31 +167,6 @@ impl C8YHttpProxyActor {
             };
         }
         info!(target: self.name(), "initialisation done.");
-    }
-
-    pub fn url_is_in_my_tenant_domain(&self, url: &str) -> bool {
-        // c8y URL may contain either `Tenant Name` or Tenant Id` so they can be one of following options:
-        // * <tenant_name>.<domain> eg: sample.c8y.io
-        // * <tenant_id>.<domain> eg: t12345.c8y.io
-        // These URLs may be both equivalent and point to the same tenant.
-        // We are going to remove that and only check if the domain is the same.
-        let tenant_uri = &self.end_point.c8y_host;
-        let url_host = match Url::parse(url) {
-            Ok(url) => match url.host() {
-                Some(host) => host.to_string(),
-                None => return false,
-            },
-            Err(_err) => {
-                return false;
-            }
-        };
-
-        let url_domain = url_host.splitn(2, '.').collect::<Vec<&str>>();
-        let tenant_domain = tenant_uri.splitn(2, '.').collect::<Vec<&str>>();
-        if url_domain.get(1) == tenant_domain.get(1) {
-            return true;
-        }
-        false
     }
 
     async fn try_get_and_set_internal_id(&mut self) -> Result<(), C8YRestError> {
@@ -336,7 +311,10 @@ impl C8YHttpProxyActor {
     async fn download_file(&mut self, request: DownloadFile) -> Result<Unit, C8YRestError> {
         let mut download_info: DownloadInfo = request.download_url.as_str().into();
         // If the provided url is c8y, add auth
-        if self.url_is_in_my_tenant_domain(download_info.url()) {
+        if self
+            .end_point
+            .url_is_in_my_tenant_domain(download_info.url())
+        {
             let token = self.get_jwt_token().await?;
             download_info.auth = Some(Auth::new_bearer(token.as_str()));
         }
@@ -347,9 +325,12 @@ impl C8YHttpProxyActor {
         downloader.download(&download_info).await?;
 
         // Move the downloaded file to the final destination
-        downloader
-            .rename(request.file_path, request.file_permissions)
-            .await?;
+        move_file(
+            downloader.filename(),
+            request.file_path,
+            request.file_permissions,
+        )
+        .await?;
 
         Ok(())
     }
