+++ crates/bin/c8y_configuration_poc/src/c8y_http_proxy/messages.rs
@@ -32,6 +32,9 @@ pub enum C8YRestError {
 
     #[error(transparent)]
     FromDownloadError(#[from] download::DownloadError),
+
+    #[error(transparent)]
+    FromFileError(#[from] tedge_utils::file::FileError),
 }
 
 pub type C8YRestResult = Result<C8YRestResponse, C8YRestError>;
