+++ crates/common/download/Cargo.toml
@@ -19,7 +19,6 @@ reqwest = { version = "0.11", default-features = false, features = [
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "1.0"
-tedge_utils = { path = "../tedge_utils" }
 tokio = { version = "1.23", features = ["fs"] }
 url = "2.2"
